Intel RealSense Camera (#358)
* Add RealSense Camera
* Send both images (Depth and RGB) with two output channels.
* Update ThermalExpert

+++ src/PlusDataCollection/InfraredTEQ1Cam/vtkInfraredTEQ1Cam.h
@@ -58,7 +58,7 @@ protected:
   int device;
 
   i3::TE_B *pTE;
-  unsigned short *pImgBuf;
+  float *pImgBuf;
   int width;
   int height;
